We are no longer interested in FIFO_CTRL/SRC...
We used to read two additional bytes after the measured
accellerations/gyro rates to look at the FIFO status. But
it seems that right after reading the fifo, these are clobbered
(they read back 0xff/0x00 randomly). So it does not make sense.

Furthermore we change how to detect an empty fifo: Instead of
checking the "EMPTY" bit, we compare fifo level to be >0. Should
yield identical results.